Include DNS query types.
It turns out that this measurement is pretty useless without this being
in it - so we add keys. We might want to limit the query types in the
future - but let's see.

We will probably also have to include the full keys and observation in
the log record to allow other uses - and then just select the used ones.
But - that is for the future.